Expose internals
