+++ .gitignore
@@ -1,4 +1,5 @@
 dist/
+dist-newstyle/
 *~
 .stack-work/
 *.hi
