+++ async.cabal
@@ -65,6 +65,7 @@ library
     if impl(ghc>=7.1)
         other-extensions: Trustworthy
     exposed-modules:     Control.Concurrent.Async
+                         Control.Concurrent.Async.Internal
     build-depends:       base     >= 4.3     && < 4.18,
                          hashable >= 1.1.2.0 && < 1.5,
                          stm      >= 2.2     && < 2.6
